+++ src/m365/aad/commands.ts
@@ -5,6 +5,7 @@ export default {
   APP_SET: `${prefix} app set`,
   APP_ROLE_ADD: `${prefix} app role add`,
   APP_ROLE_LIST: `${prefix} app role list`,
+  APP_ROLE_DELETE: `${prefix} app role delete`,
   APPROLEASSIGNMENT_ADD: `${prefix} approleassignment add`,
   APPROLEASSIGNMENT_LIST: `${prefix} approleassignment list`,
   APPROLEASSIGNMENT_REMOVE: `${prefix} approleassignment remove`,