+++ docs/docs/about/release-notes.md
@@ -1,11 +1,34 @@
 # Release notes
 
+## [v3.12.0](https://github.com/pnp/cli-microsoft365/releases/tag/v3.12.0)
+
+### New commands
+
+**OneDrive:**
+
+- [onedrive list](../cmd/onedrive/onedrive-list.md) - retrieves a list of OneDrive sites [#2230](https://github.com/pnp/cli-microsoft365/issues/2230)
+
+**Planner:**
+
+- [planner bucket add](../cmd/planner/bucket/bucket-add.md) - adds a new Microsoft Planner bucket [#2254](https://github.com/pnp/cli-microsoft365/issues/2254)
+- [planner bucket list](../cmd/planner/bucket/bucket-list.md) - lists the Microsoft Planner buckets in a plan [#2556](https://github.com/pnp/cli-microsoft365/issues/2556)
+- [planner plan list](../cmd/planner/plan/plan-list.md) - returns a list of plans associated with a specified group [#2499](https://github.com/pnp/cli-microsoft365/issues/2499)
+
+### Changes
+
+- added the 'Add multiple lists in multiple sites' sample [#2562](https://github.com/pnp/cli-microsoft365/pull/2562)
+- updated 'teams user app list' to Graph v1.0 [#2510](https://github.com/pnp/cli-microsoft365/issues/2510)
+- added the 'List failed site designs' sample [#2539](https://github.com/pnp/cli-microsoft365/issues/2539)
+- extended 'spo web get' with groups [#2493](https://github.com/pnp/cli-microsoft365/issues/2493)
+- fixed 'yammer message add' [#2557](https://github.com/pnp/cli-microsoft365/issues/2557)
+
 ## [v3.11.0](https://github.com/pnp/cli-microsoft365/releases/tag/v3.11.0)
 
 ### New commands
 
 **Azure Active Directory:**
 
+- [aad app role delete](../cmd/aad/app/app-role-delete.md) - deletes role from the specified Azure AD app registration [#2150](https://github.com/pnp/cli-microsoft365/issues/2150)
 - [aad app set](../cmd/aad/app/app-set.md) - updates Azure AD app registration [#2448](https://github.com/pnp/cli-microsoft365/issues/2448)
 - [aad o365group recyclebinitem clear](../cmd/aad/o365group/o365group-recyclebinitem-clear.md) - clears Microsoft 365 Groups from the recycle bin in the current tenant [#2033](https://github.com/pnp/cli-microsoft365/issues/2033)